feat: o2-76 auto-add mermaid metadata tag if markdown file contains mermaid syntax block
